saltbox/requirements.txt
GitPYthon
github3.py
plugnparse
salt
docker
